Check if drush dir exists.

+++ Vagrantfile
@@ -45,19 +45,21 @@ if vconfig['create_drush_alias']
     end
   end
 
-  alias_file = "#{Dir.home}/.drush/"+vconfig['vagrant_hostname']+".aliases.drushrc.php"
-  if ARGV[0] == "destroy"
-    File.delete(alias_file) if File.exist?(alias_file)
-  else
-     alias_file = File.open(alias_file, "w+")
-     template = File.read("./vagrant-includes/drush_alias.erb")
-     da = DrushAlias.new
-     da.hostname = vconfig['vagrant_hostname']
-     da.uri = vconfig['apache_vhosts'][0]['servername']
-     da.ip = vconfig['vagrant_ip']
-     da.root = vconfig['vagrant_synced_folders'][0]['destination']
-     alias_file << ERB.new(template).result(da.template_binding)
-     alias_file.close
+  if File.directory?("#{Dir.home}/.drush")
+    alias_file = "#{Dir.home}/.drush/"+vconfig['vagrant_hostname']+".aliases.drushrc.php"
+    if ARGV[0] == "destroy"
+      File.delete(alias_file) if File.exist?(alias_file)
+    else
+       alias_file = File.open(alias_file, "w+")
+       template = File.read("./vagrant-includes/drush_alias.erb")
+       da = DrushAlias.new
+       da.hostname = vconfig['vagrant_hostname']
+       da.uri = vconfig['apache_vhosts'][0]['servername']
+       da.ip = vconfig['vagrant_ip']
+       da.root = vconfig['vagrant_synced_folders'][0]['destination']
+       alias_file << ERB.new(template).result(da.template_binding)
+       alias_file.close
+    end
   end
 
 end
